+++ src/main/java/com/grupoy/tpbd2/model/Producto.java
@@ -1,5 +1,7 @@
 package com.grupoy.tpbd2.model;
 
+import java.math.BigDecimal;
+
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
@@ -14,5 +16,5 @@ public class Producto {
     private String descripcion;
     private String tipo;
     private String laboratorio;
-    private Double precioActual;
+    private BigDecimal precioActual;
 }
